+++ src/main/java/com/ametrinstudios/ametrin/mixin/MixinEnchantment.java
@@ -1,13 +1,13 @@
 package com.ametrinstudios.ametrin.mixin;
 
-import com.ametrinstudios.ametrin.mixin.util.IEnchantmentMixin;
+import com.ametrinstudios.ametrin.util.mixin.IMixinEnchantment;
 import net.minecraft.world.entity.EquipmentSlot;
 import net.minecraft.world.item.enchantment.Enchantment;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
 
 @Mixin(Enchantment.class)
-public abstract class EnchantmentMixin implements IEnchantmentMixin {
+public abstract class MixinEnchantment implements IMixinEnchantment {
     @Shadow EquipmentSlot[] slots;
 
     @Override