+++ src/main/java/com/ametrinstudios/ametrin/util/mixin/IMixinEnchantment.java
@@ -1,7 +1,9 @@
 package com.ametrinstudios.ametrin.util.mixin;
 
 import net.minecraft.world.entity.EquipmentSlot;
+import net.minecraft.world.entity.LivingEntity;
 
 public interface IMixinEnchantment {
     EquipmentSlot[] getApplicableSlots();
+    boolean hasEnchantment(LivingEntity entity);
 }