+++ src/main/java/com/ametrinstudios/ametrin/world/block/FlammableBlock.java
@@ -16,10 +16,7 @@ public class FlammableBlock extends Block {
         this.fireSpreadSpeed = fireSpreadSpeed;
     }
 
-    @Override
-    public boolean isFlammable(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return true;}
-    @Override
-    public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return fireSpreadSpeed;}
-    @Override
-    public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return fireSpreadSpeed;}
+    @Override public boolean isFlammable(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return true;}
+    @Override public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return fireSpreadSpeed;}
+    @Override public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return fireSpreadSpeed;}
 }