+++ src/main/java/com/ametrinstudios/ametrin/world/block/FlammableLogBlock.java
@@ -9,10 +9,7 @@ import net.minecraft.world.level.block.state.BlockState;
 public class FlammableLogBlock extends RotatedPillarBlock {
     public FlammableLogBlock(Properties properties) {super(properties);}
 
-    @Override
-    public boolean isFlammable(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return true;}
-    @Override
-    public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 5;}
-    @Override
-    public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 5;}
+    @Override public boolean isFlammable(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return true;}
+    @Override public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 5;}
+    @Override public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 5;}
 }