+++ src/main/java/com/ametrinstudios/ametrin/world/block/FlammableStairPlankBlock.java
@@ -11,10 +11,7 @@ import java.util.function.Supplier;
 public class FlammableStairPlankBlock extends StairBlock {
     public FlammableStairPlankBlock(Supplier<BlockState> modelState, Properties properties) {super(modelState, properties);}
 
-    @Override
-    public boolean isFlammable(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return true;}
-    @Override
-    public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 5;}
-    @Override
-    public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 20;}
+    @Override public boolean isFlammable(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return true;}
+    @Override public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 5;}
+    @Override public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 20;}
 }