+++ src/main/java/com/ametrinstudios/ametrin/world/block/SimpleDoublePlantBlock.java
@@ -24,10 +24,7 @@ public class SimpleDoublePlantBlock extends DoublePlantBlock {
         return blockState.getValue(HALF) == DoubleBlockHalf.UPPER ? UpperShape : LowerShape;
     }
 
-    @Override
-    public boolean isFlammable(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return true;}
-    @Override
-    public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 60;}
-    @Override
-    public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 100;}
+    @Override public boolean isFlammable(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return true;}
+    @Override public int getFlammability(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 60;}
+    @Override public int getFireSpreadSpeed(BlockState state, BlockGetter level, BlockPos pos, Direction direction) {return 100;}
 }