+++ changelog.md
@@ -1,9 +1,9 @@
 ## 0.0.26 (ALPHA)
 * added `getApplicableSlots()` to `Enchantment`
-* fixed `BlockBehaviorPropertiesMixin`
 * added more flammable blocks
 * renamed `FlammableRotatedPillarBlock` to `FlammableLogBlock`
 * `LiquidBlock`s now generate an empty model with particle texture
+* fixes mixins
 
 ## 0.0.25 (ALPHA)
 * added `FlammableBlock` and default implementations for Leaves, Planks and Logs/Woods